Big Progress and Future alert
After days of producing styles and running multiple tests, this version is close to being finished. Now, the backend needs to be implemented to make Digimon data easier to find and clearer. In the future, a section for team testing (weaknesses and strengths) will also be added.

My friends got on my nerves for not putting the process here, so I just put where I am from and where I'm going.